Title: 최소 에너지 비용, Time: 243ms, Memory: 16MB, Status: Wrong Answer - Codetree

+++ 250528/최소 에너지 비용/minimum-energy-cost.java
@@ -17,16 +17,16 @@ public class Main {
         }
 
         long min = (long)Integer.MAX_VALUE;
-        int total = 0;
+        long total = 0;
         for (int i = 0; i<n; i++) {
             
             if (i == 0) {
                 total = 0;
             }
             else {
-                total += dist[i-1] * cost[i-1];
+                total += (long)dist[i-1] * cost[i-1];
             }
-            min = Math.min(min, (long)total + cost[i] * R[i]);
+            min = Math.min(min, total + (long)cost[i] * R[i]);
         }
         
         System.out.print(min);
